# BackendSpring/src/main/java/heg/backendspring/api/CourseApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.9.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package heg.backendspring.api;

import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-11-23T19:52:08.287321300+01:00[Europe/Zurich]", comments = "Generator version: 7.9.0")
@Validated
@Tag(name = "course", description = "Consultation des cours et opérations associées")
public interface CourseApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /course/{courseId} : Détails d’un cours.
     * Retourne les informations d’un cours spécifique.
     *
     * @param courseId Identifiant du cours. (required)
     * @return Cours trouvé. (status code 200)
     *         or Cours introuvable. (status code 404)
     *         or Erreur interne. (status code 500)
     */
    @Operation(
        operationId = "getCourseById",
        summary = "Détails d’un cours.",
        description = "Retourne les informations d’un cours spécifique.",
        tags = { "course" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Cours trouvé.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = heg.backendspring.models.CourseDto.class))
            }),
            @ApiResponse(responseCode = "404", description = "Cours introuvable."),
            @ApiResponse(responseCode = "500", description = "Erreur interne.")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/course/{courseId}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<heg.backendspring.models.CourseDto> getCourseById(
        @Parameter(name = "courseId", description = "Identifiant du cours.", required = true, in = ParameterIn.PATH) @PathVariable("courseId") Long courseId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"id\" : 1, \"name\" : \"Python Introduction\", \"code\" : \"61-13\", \"semester\" : \"AUTOMNE\", \"courseType\" : \"PYTHON\", \"year_course\" : 2025, \"teacher\" : \"Stettler\", \"students\" : [ ], \"tps\" : [ { \"id\" : 10, \"false\" : 1, \"courseId\" : 1, \"submission\" : { \"id\" : 100, \"fileName\" : \"TP1_Rendus.zip\", \"pathStorage\" : \"/data/courses/61-13/TP1/TP1_Rendus.zip\", \"pathFileStructured\" : \"/data/courses/61-13/TP1/TP1_RenduRestructuration.zip\" }, \"statusStudents\" : [ ] } ], \"evaluations\" : [ { \"id\" : 200, \"name\" : \"Examen final\", \"date\" : \"2025-01-15\", \"coefficient\" : 2, \"courseId\" : 1 } ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /course : Liste des cours.
     * Retourne tous les cours accessibles.
     *
     * @return Liste de cours retournée. (status code 200)
     *         or Erreur interne. (status code 500)
     */
    @Operation(
        operationId = "getCourses",
        summary = "Liste des cours.",
        description = "Retourne tous les cours accessibles.",
        tags = { "course" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Liste de cours retournée.", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = heg.backendspring.models.CourseDto.class)))
            }),
            @ApiResponse(responseCode = "500", description = "Erreur interne.")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/course",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<heg.backendspring.models.CourseDto>> getCourses(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"id\" : 1, \"name\" : \"Python Introduction\", \"code\" : \"61-13\", \"semester\" : \"AUTOMNE\", \"courseType\" : \"PYTHON\", \"year_course\" : 2025, \"teacher\" : \"Stettler\", \"students\" : [ ], \"tps\" : [ { \"id\" : 10, \"false\" : 1, \"courseId\" : 1, \"submission\" : { \"id\" : 100, \"fileName\" : \"TP1_Rendus.zip\", \"pathStorage\" : \"/data/courses/61-13/TP1/TP1_Rendus.zip\", \"pathFileStructured\" : \"/data/courses/61-13/TP1/TP1_RenduRestructuration.zip\" }, \"statusStudents\" : [ ] } ], \"evaluations\" : [ { \"id\" : 200, \"name\" : \"Examen final\", \"date\" : \"2025-01-15\", \"coefficient\" : 2, \"courseId\" : 1 } ] }, { \"id\" : 1, \"name\" : \"Python Introduction\", \"code\" : \"61-13\", \"semester\" : \"AUTOMNE\", \"courseType\" : \"PYTHON\", \"year_course\" : 2025, \"teacher\" : \"Stettler\", \"students\" : [ ], \"tps\" : [ { \"id\" : 10, \"false\" : 1, \"courseId\" : 1, \"submission\" : { \"id\" : 100, \"fileName\" : \"TP1_Rendus.zip\", \"pathStorage\" : \"/data/courses/61-13/TP1/TP1_Rendus.zip\", \"pathFileStructured\" : \"/data/courses/61-13/TP1/TP1_RenduRestructuration.zip\" }, \"statusStudents\" : [ ] } ], \"evaluations\" : [ { \"id\" : 200, \"name\" : \"Examen final\", \"date\" : \"2025-01-15\", \"coefficient\" : 2, \"courseId\" : 1 } ] } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}