# BackendSpring/src/main/resources/openapi/openapi.yaml
...
openapi: 3.0.4
info:
  title: AssistantIO API
  description: API de gestion des cours, TPs, étudiants et rendus.
  version: 1.0.0

# ==============================
#   SERVERS
# ==============================
servers:
  - url: http://localhost:8080
    description: Environnement local

# ==============================
#   TAGS
# ==============================
...
paths:
...
  /course:
    get:
      tags: [course]
      summary: Liste des cours.
      description: Retourne tous les cours accessibles.
      operationId: getCourses
      responses:
        '200':
          description: Liste de cours retournée.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CourseDto'
        '500':
          description: Erreur interne.
  /course/{courseId}:
    get:
      tags: [course]
      summary: Détails d’un cours.
      description: Retourne les informations d’un cours spécifique.
      operationId: getCourseById
      parameters:
        - name: courseId
          in: path
          required: true
          description: Identifiant du cours.
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Cours trouvé.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseDto'
        '404':
          description: Cours introuvable.
        '500':
          description: Erreur interne.

  # ========================================
  #   STUDENTS D’UN COURS (REST PRINCIPAL)
  # ========================================
...
components:
  schemas:
...
    SemesterType:
      type: string
      description: Semestre académique.
      example: "AUTOMNE"
    CourseType:
      type: string
      description: Type de cours.
      example: "PYTHON"
    StudyType:
      type: string
      description: Type d’étude.
      example: "FULL_TIME"

    # --- DTO principaux exposés ---
    CourseDto:
      type: object
      description: Représentation d’un cours.
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "Python Introduction"
        code:
          type: string
          example: "61-13"
        semester:
          $ref: '#/components/schemas/SemesterType'
        courseType:
          $ref: '#/components/schemas/CourseType'
        year_course:
          type: integer
          example: 2025
        teacher:
          type: string
          example: "Stettler"
        students:
          type: array
          items:
            $ref: '#/components/schemas/StudentDto'
        tps:
          type: array
          items:
            $ref: '#/components/schemas/TPDto'
        evaluations:
          type: array
          items:
            $ref: '#/components/schemas/EvaluationDto'
      required: ["id", "name", "code"]
      example:
        id: 1
        name: "Python Introduction"
        code: "61-13"
        semester: "AUTOMNE"
        courseType: "PYTHON"
        year_course: 2025
        teacher: "Stettler"
        students: []
        tps:
          - id: 10
            no: 1
            courseId: 1
            submission:
              id: 100
              fileName: "TP1_Rendus.zip"
              pathStorage: "/data/courses/61-13/TP1/TP1_Rendus.zip"
              pathFileStructured: "/data/courses/61-13/TP1/TP1_RenduRestructuration.zip"
            statusStudents: []
        evaluations:
          - id: 200
            name: "Examen final"
            date: "2025-01-15"
            coefficient: 2
            courseId: 1
    EvaluationDto:
      type: object
      description: Représentation d’une évaluation.
      properties:
        id:
          type: integer
          format: int64
          example: 200
        name:
          type: string
          example: "Examen final"
        date:
          type: string
          format: date
          example: "2025-01-15"
        coefficient:
          type: integer
          example: 2
        courseId:
          type: integer
          format: int64
          example: 1
      required: ["name", "date", "coefficient", "courseId"]
      example:
        id: 200
        name: "Examen final"
        date: "2025-01-15"
        coefficient: 2
        courseId: 1
    StudentDto:
      type: object
      description: Représentation d’un étudiant.
      properties:
        id:
          type: integer
          format: int64
          example: 300
        name:
          type: string
          example: "George Dylan"
        email:
          type: string
          format: email
          example: "george.dylan@etu.hesge.ch"
        studyType:
          $ref: '#/components/schemas/StudyType'
        studentCourses:
          type: array
          items:
            $ref: '#/components/schemas/CourseDto'
      required: ["name", "email", "studyType"]
      example:
        id: 300
        name: "George Dylan"
        email: "george.dylan@etu.hesge.ch"
        studyType: "FULL_TIME"
        studentCourses:
          - id: 1
            name: "Python Introduction"
            code: "61-13"
            semester: "AUTOMNE"
            courseType: "PYTHON"
            year_course: 2025
            teacher: "Stettler"
    SubmissionDto:
      type: object
      description: Représentation d’un rendu (fichier ZIP) pour un TP.
      properties:
        id:
          type: integer
          format: int64
          example: 100
        fileName:
          type: string
          example: "TP1_Rendus.zip"
        pathStorage:
          type: string
          description: Chemin de stockage du ZIP original.
          example: "/data/courses/61-13/TP1/TP1_Rendus.zip"
        pathFileStructured:
          type: string
          description: Chemin du ZIP restructuré.
          example: "/data/courses/61-13/TP1/TP1_RenduRestructuration.zip"
      required: ["fileName", "pathStorage"]
      example:
        id: 100
        fileName: "TP1_Rendus.zip"
        pathStorage: "/data/courses/61-13/TP1/TP1_Rendus.zip"
        pathFileStructured: "/data/courses/61-13/TP1/TP1_RenduRestructuration.zip"
    TPDto:
      type: object
      description: Représentation d’un TP pour un cours.
      properties:
        id:
          type: integer
          format: int64
          example: 10
        "no":
          type: integer
          example: 1
        courseId:
          type: integer
          format: int64
          example: 1
        submission:
          $ref: '#/components/schemas/SubmissionDto'
        statusStudents:
          type: array
          items:
            $ref: '#/components/schemas/TPStatusDto'
      required: ["no", "courseId"]
      example:
        id: 10
        no: 1
        courseId: 1
        submission:
          id: 100
          fileName: "TP1_Rendus.zip"
          pathStorage: "/data/courses/61-13/TP1/TP1_Rendus.zip"
          pathFileStructured: "/data/courses/61-13/TP1/TP1_RenduRestructuration.zip"
        statusStudents:
          - id: 1000
            studentId: 300
            tpId: 10
            studentSubmission: true
    TPStatusDto:
      type: object
      description: Statut d’un étudiant pour un TP.
      properties:
        id:
          type: integer
          format: int64
          example: 1000
        studentId:
          type: integer
          format: int64
          example: 300
        tpId:
          type: integer
          format: int64
          example: 10
        studentSubmission:
          type: boolean
          example: true
      required: ["studentId", "tpId", "studentSubmission"]
      example:
        id: 1000
        studentId: 300
        tpId: 10
        studentSubmission: true